edge_vision_defect_detector/pyproject.toml
[tool.poetry]
name = "edge-vision-defect-detector"
version = "0.1.0"
description = "YOLOv8 + RL-assisted defect detection deployed on NVIDIA Jetson Orin"
authors = ["Hari <youremail@example.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{include = "edge_detector", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.2"
torchvision = "^0.17"
ultralytics = "^8.2"           # YOLOv8
pytorch-lightning = "^2.2"
opencv-python = "^4.10"
onnx = "^1.16"
mlflow = "^2.12"
tensorrt = {version = "^10.0", optional = true}
kafka-python = "^2.0"
fastapi = "^0.111"
pydantic = "^2.7"
faiss-cpu = "^1.8"
opentelemetry-sdk = "^1.25"
prometheus-client = "^0.20"
wandb = "^0.18"
gymnasium = "^0.29"
stable-baselines3 = {extras = ["extra"], version = "^2.3"}
numpy = ">=1.24,<2.0"

[tool.poetry.group.dev.dependencies]
black = "^24.3"
ruff = "^0.4"
pytest = "^8.2"
pytest-benchmark = "^4.0"
pre-commit = "^3.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
